+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/VideoPlayerOptions.java
@@ -6,4 +6,6 @@ package io.flutter.plugins.videoplayer;
 
 public class VideoPlayerOptions {
   public boolean mixWithOthers;
+  public boolean enableDecoderFallback;
+  public boolean disableMediaCodecAsyncQueueing;
 }